task 6 corrections

+++ 0x00-hello_world/6-size.c
@@ -5,10 +5,10 @@
 */
 int main(void)
 {
-printf("Size of char: %ld byte\n", sizeof(char));
-printf("Size of int: %ld bytes\n", sizeof(int));
-printf("Size of long int: %ld byte\n", sizeof(long int));
-printf("Size of long long int: %ld bytes\n", sizeof(long long int));
-printf("Size of float: %ld bytes\n", sizeof(float));
+printf("Size of char: %ld byte\n", (unsigned long)sizeof(char));
+printf("Size of int: %ld bytes\n", (unsigned long)sizeof(int));
+printf("Size of long int: %ld byte\n", (unsigned long)sizeof(long int));
+printf("Size of long long int: %ld bytes\n", (unsigned long)sizeof(long long));
+printf("Size of float: %ld bytes\n", (unsigned long)sizeof(float));
 return (0);
 }
